requirements.txt
streamlit
torch
torchvision
timm
faiss-cpu
pandas
Pillow
requests
tqdm
numpy
fastapi
uvicorn